requirements.txt
streamlit
python-docx
together
